client/test: Reduce setup complexity
Previously, we introduced error channels for immediate error handling.
However, this was done at the expense of requiring an additional error
channel parameter during testing setup, which made it more complicated
to setup a client test. Now the error channels are kept only with the
role and the changes to the setup function parameters are reverted.